fast_api/requirements.txt
fastapi
uvicorn[standard]
google-cloud-translate
google-auth
python-dotenv
pymongo
bcrypt
python-jose[cryptography]
passlib[bcrypt]
numpy